Remplacement de `dealWithProcols` par `defaillances`
